Move GroupAdminAction to its own module outside admin package, as it is not an admin action
